services/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.34
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic-settings==2.4.0
alembic==1.13.2
redis==5.0.7
celery==5.4.0
boto3==1.34.161
httpx==0.27.2
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cryptography==41.0.7
svix==1.18.0
sse-starlette==1.8.2
sentry-sdk[fastapi]==1.38.0
prometheus-fastapi-instrumentator==6.1.0
python-json-logger==2.0.7
ruff==0.6.9
pytest==8.3.2
pytest-cov==4.1.0
